+++ libs/angular/src/lib/green-angular.module.ts
@@ -12,6 +12,8 @@ import { NggProgressCircleModule } from './progress-circle/progress-circle.modul
 import { NggSegmentedControlModule } from './segmented-control/segmented-control.module'
 import { NggSharedModule } from './shared/shared.module'
 import { NggSliderModule } from './slider/slider.module'
+import { NggSortableListModule } from './sortable-list/sortable-list.module'
+import { NggCellTableModule } from './cell-table'
 
 @NgModule({
   declarations: [],
@@ -28,7 +30,9 @@ import { NggSliderModule } from './slider/slider.module'
     NggSliderModule,
     NggContextMenuModule,
     NggInPageWizardModule,
+    NggCellTableModule,
     NggSharedModule,
+    NggSortableListModule,
   ],
 })
 export class NggModule {}