+++ libs/core/src/components/datepicker/datepicker.test.ts
@@ -176,6 +176,38 @@ describe('<gds-datepicker>', () => {
 
       await expect(focusedDate).to.be.undefined
     })
+
+    it('Setting `disabled-weekends` should disable weekends', async () => {
+      const el = await fixture<GdsDatepicker>(
+        html`<gds-datepicker
+          value="2024-01-10"
+          disabled-weekends
+          open
+        ></gds-datepicker>`
+      )
+
+      await el.updateComplete
+
+      const disabledDatecell = await el.test_getDateCell(13)
+
+      await expect(disabledDatecell).to.have.attribute('disabled')
+    })
+
+    it('Setting `disabled-dates` should disable dates', async () => {
+      const el = await fixture<GdsDatepicker>(
+        html`<gds-datepicker
+          value="2024-01-10"
+          disabled-dates="2024-01-13"
+          open
+        ></gds-datepicker>`
+      )
+
+      await el.updateComplete
+
+      const disabledDatecell = await el.test_getDateCell(13)
+
+      await expect(disabledDatecell).to.have.attribute('disabled')
+    })
   })
 
   describe('Interactions', () => {
@@ -361,7 +393,7 @@ describe('<gds-datepicker>', () => {
       const popover =
         el.shadowRoot!.querySelector<GdsPopover>('#calendar-popover')!
       const monthDropdown = el.shadowRoot!.querySelector<GdsDropdown>(
-        `${getScopedTagName('gds-dropdown')}[aria-label="Month"]`
+        `${getScopedTagName('gds-dropdown')}[label="Month"]`
       )!
 
       await clickOnElement(button)
@@ -429,7 +461,7 @@ describe('<gds-datepicker>', () => {
 
     it('should give calendar keyboard focus after opening the popover', async () => {
       const el = await fixture<GdsDatepicker>(
-        html`<gds-datepicker></gds-datepicker>`
+        html`<gds-datepicker value="2024-01-01"></gds-datepicker>`
       )
 
       const button = el.shadowRoot!.querySelector<HTMLButtonElement>(
@@ -445,7 +477,7 @@ describe('<gds-datepicker>', () => {
         press: 'Enter',
       })
 
-      expect(onlyDate(el.value!)).to.equal(onlyDate(new Date()))
+      expect(onlyDate(el.value!)).to.equal(onlyDate(new Date('2024-01-01')))
     })
 
     it('should set spinners to yyyy, mm and dd when date is undefined', async () => {